+++ tests/UT_Money.cpp
@@ -112,14 +112,14 @@ namespace fmc {
     TEST(Test_Money, multiplication) {
         {
             Money m{42};
-            Money m2 = m * 2;
+            Money m2 = m * (long long int) 2;
 
             EXPECT_EQ(m2.dollars, 84);
             EXPECT_EQ(m2.cents, 0);
         }
         {
             Money m{42.5};
-            Money m2 = m * 2;
+            Money m2 = m * (long long int) 2;
 
             EXPECT_EQ(m2.dollars, 85);
             EXPECT_EQ(m2.cents, 0);
@@ -127,7 +127,7 @@ namespace fmc {
         {
             Money m{42, 30};
             EXPECT_EQ(m.cents, 30);
-            Money m2 = m * 2;
+            Money m2 = m * (long long int) 2;
 
             EXPECT_EQ(m2.dollars, 84);
             EXPECT_EQ(m2.cents, 60);
